PEER-214: Clean up utils code
