+++ backend/collaboration/src/lib/y-postgres/constants.ts
@@ -1,0 +1,10 @@
+export const wsReadyStateConnecting = 0;
+export const wsReadyStateOpen = 1;
+export const wsReadyStateClosing = 2; // eslint-disable-line
+export const wsReadyStateClosed = 3; // eslint-disable-line
+
+export const messageSync = 0;
+export const messageAwareness = 1;
+// const messageAuth = 2
+
+export const pingTimeout = 30000;
