+++ backend/collaboration/src/lib/y-postgres/utils.ts
@@ -1,24 +1,20 @@
-/*
-Implementation from yjs-websocket/bin/utils.js
-Copied here for safe referencing
-*/
-import * as Y from 'yjs';
-import * as syncProtocol from 'y-protocols/sync';
 import * as awarenessProtocol from 'y-protocols/awareness';
+import * as syncProtocol from 'y-protocols/sync';
 
-import * as encoding from 'lib0/encoding';
 import * as decoding from 'lib0/decoding';
+import * as encoding from 'lib0/encoding';
 import * as map from 'lib0/map';
 
-import { IPersistence, IWSSharedDoc } from '@/types/interfaces';
+import type { IPersistence, IWSSharedDoc } from '@/types/interfaces';
 
-const wsReadyStateConnecting = 0;
-const wsReadyStateOpen = 1;
-const wsReadyStateClosing = 2; // eslint-disable-line
-const wsReadyStateClosed = 3; // eslint-disable-line
-
-// disable gc when using snapshots!
-const gcEnabled = process.env.GC !== 'false' && process.env.GC !== '0';
+import {
+  messageAwareness,
+  messageSync,
+  pingTimeout,
+  wsReadyStateConnecting,
+  wsReadyStateOpen,
+} from './constants';
+import { WSSharedDoc } from './ws-shared-doc';
 
 let persistence: IPersistence | null = null;
 
@@ -30,73 +26,6 @@ export const getPersistence = () => persistence;
 
 // exporting docs so that others can use it
 export const docs = new Map<string, IWSSharedDoc>();
-
-const messageSync = 0;
-const messageAwareness = 1;
-// const messageAuth = 2
-
-const updateHandler = (update: Uint8Array, origin: any, doc: Y.Doc) => {
-  const sharedDoc = doc as IWSSharedDoc;
-
-  const encoder = encoding.createEncoder();
-  encoding.writeVarUint(encoder, messageSync);
-  syncProtocol.writeUpdate(encoder, update);
-  const message = encoding.toUint8Array(encoder);
-  sharedDoc.conns.forEach((_, conn) => send(sharedDoc, conn, message));
-};
-
-class WSSharedDoc extends Y.Doc implements IWSSharedDoc {
-  name: string;
-  conns: Map<object, Set<number>>;
-  awareness: awarenessProtocol.Awareness;
-
-  constructor(name: string) {
-    super({ gc: gcEnabled });
-    this.name = name;
-    this.conns = new Map();
-    this.awareness = new awarenessProtocol.Awareness(this);
-    this.awareness.setLocalState(null);
-
-    const awarenessChangeHandler = (
-      {
-        added,
-        updated,
-        removed,
-      }: {
-        added: Array<number>;
-        updated: Array<number>;
-        removed: Array<number>;
-      },
-      conn: object | null
-    ) => {
-      const changedClients = added.concat(updated, removed);
-      if (conn !== null) {
-        const connControlledIDs = /** @type {Set<number>} */ this.conns.get(conn);
-        if (connControlledIDs !== undefined) {
-          added.forEach((clientID) => {
-            connControlledIDs.add(clientID);
-          });
-          removed.forEach((clientID) => {
-            connControlledIDs.delete(clientID);
-          });
-        }
-      }
-      // broadcast awareness update
-      const encoder = encoding.createEncoder();
-      encoding.writeVarUint(encoder, messageAwareness);
-      encoding.writeVarUint8Array(
-        encoder,
-        awarenessProtocol.encodeAwarenessUpdate(this.awareness, changedClients)
-      );
-      const buff = encoding.toUint8Array(encoder);
-      this.conns.forEach((_, c) => {
-        send(this, c, buff);
-      });
-    };
-    this.awareness.on('update', awarenessChangeHandler);
-    this.on('update', updateHandler);
-  }
-}
 
 /**
  * Gets a Y.Doc by name, whether in memory or on disk
@@ -165,13 +94,13 @@ const closeConn = (doc: IWSSharedDoc, conn: any) => {
   conn.close();
 };
 
-const send = (doc: IWSSharedDoc, conn: any, m: Uint8Array) => {
+export const send = (doc: IWSSharedDoc, conn: any, m: Uint8Array) => {
   if (conn.readyState !== wsReadyStateConnecting && conn.readyState !== wsReadyStateOpen) {
     closeConn(doc, conn);
   }
   try {
     conn.send(m, (err: any) => {
-      if (err != null) {
+      if (err !== null) {
         closeConn(doc, conn);
       }
     });
@@ -179,8 +108,6 @@ const send = (doc: IWSSharedDoc, conn: any, m: Uint8Array) => {
     closeConn(doc, conn);
   }
 };
-
-const pingTimeout = 30000;
 
 export const setupWSConnection = (
   conn: any,
