+++ backend/collaboration/src/lib/y-postgres/ws-shared-doc.ts
@@ -1,0 +1,73 @@
+import * as Y from 'yjs';
+import * as syncProtocol from 'y-protocols/sync';
+import * as awarenessProtocol from 'y-protocols/awareness';
+
+import * as encoding from 'lib0/encoding';
+
+import type { IWSSharedDoc } from '@/types/interfaces';
+import { GC_ENABLED } from '@/config';
+import { messageAwareness, messageSync } from './constants';
+import { send } from './utils';
+
+const updateHandler = (update: Uint8Array, _origin: any, doc: Y.Doc) => {
+  const sharedDoc = doc as IWSSharedDoc;
+
+  const encoder = encoding.createEncoder();
+  encoding.writeVarUint(encoder, messageSync);
+  syncProtocol.writeUpdate(encoder, update);
+  const message = encoding.toUint8Array(encoder);
+  sharedDoc.conns.forEach((_, conn) => send(sharedDoc, conn, message));
+};
+
+export class WSSharedDoc extends Y.Doc implements IWSSharedDoc {
+  name: string;
+  conns: Map<object, Set<number>>;
+  awareness: awarenessProtocol.Awareness;
+
+  constructor(name: string) {
+    super({ gc: GC_ENABLED });
+    this.name = name;
+    this.conns = new Map();
+    this.awareness = new awarenessProtocol.Awareness(this);
+    this.awareness.setLocalState(null);
+
+    const awarenessChangeHandler = (
+      {
+        added,
+        updated,
+        removed,
+      }: {
+        added: Array<number>;
+        updated: Array<number>;
+        removed: Array<number>;
+      },
+      conn: object | null
+    ) => {
+      const changedClients = added.concat(updated, removed);
+      if (conn !== null) {
+        const connControlledIDs = /** @type {Set<number>} */ this.conns.get(conn);
+        if (connControlledIDs !== undefined) {
+          added.forEach((clientID) => {
+            connControlledIDs.add(clientID);
+          });
+          removed.forEach((clientID) => {
+            connControlledIDs.delete(clientID);
+          });
+        }
+      }
+      // broadcast awareness update
+      const encoder = encoding.createEncoder();
+      encoding.writeVarUint(encoder, messageAwareness);
+      encoding.writeVarUint8Array(
+        encoder,
+        awarenessProtocol.encodeAwarenessUpdate(this.awareness, changedClients)
+      );
+      const buff = encoding.toUint8Array(encoder);
+      this.conns.forEach((_, c) => {
+        send(this, c, buff);
+      });
+    };
+    this.awareness.on('update', awarenessChangeHandler);
+    this.on('update', updateHandler);
+  }
+}
