+++ backend/collaboration/src/server.ts
@@ -1,6 +1,7 @@
 import { exit } from 'process';
 
 import cors from 'cors';
+import http from 'http';
 import express, { json } from 'express';
 import { StatusCodes } from 'http-status-codes';
 import pino from 'pino-http';
@@ -9,13 +10,9 @@ import { UI_HOST } from '@/config';
 import { config, db } from '@/lib/db';
 import { logger } from '@/lib/utils';
 import roomRoutes from '@/routes/room';
+import { setUpWSServer } from './ws';
 
-import { WebSocketServer } from 'ws';
-import { createServer } from 'http';
-import { setupWSConnection } from './y-postgresql-util/utils';
-import { setUpPersistence } from './y-postgresql-util/persistence';
 const app = express();
-const server = createServer(app);
 
 app.use(pino());
 app.use(json());
@@ -31,10 +28,6 @@ app.use('/room', roomRoutes);
 // Health Check for Docker
 app.get('/health', (_req, res) => res.status(StatusCodes.OK).send('OK'));
 
-// y-websocket server
-const wss = new WebSocketServer({ server });
-wss.on('connection', setupWSConnection);
-setUpPersistence();
 export const dbHealthCheck = async () => {
   try {
     await db`SELECT 1`;
@@ -52,5 +45,9 @@ app.get('/test-db', async (_req, res) => {
   await dbHealthCheck();
   res.json({ message: 'OK ' });
 });
+
+const server = http.createServer(app);
+
+export const wss = setUpWSServer(server);
 
 export default server;
